requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
bio==1.8.1
biopython==1.86
biothings_client==0.4.1
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
coloredlogs==15.0.1
cryptography==46.0.3
distro==1.9.0
exceptiongroup==1.3.0
flatbuffers==25.9.23
frozenlist==1.8.0
google-auth==2.43.0
google-genai==1.49.0
gprofiler-official==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
humanfriendly==10.0
idna==3.11
jiter==0.12.0
ml_collections==1.1.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mygene==3.2.2
networkx==3.5
numpy==2.3.4
onnxruntime==1.23.2
openai==2.7.2
anthropic==0.37.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
propcache==0.4.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyMuPDF==1.26.6
pymupdf-layout==1.26.6
pymupdf4llm==0.2.0
pypdfium2==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tqdm==4.67.1
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
